Rename LoginController file
